requirements.txt
# Wake word detection
pvporcupine>=3.0.0

# Audio processing
pyaudio>=0.2.11
sounddevice>=0.4.6
numpy>=1.24.0

# WebSocket communication
websockets>=12.0
certifi>=2023.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Device authentication
cryptography>=41.0.0
requests>=2.31.0

# Async HTTP client for location and weather
aiohttp>=3.9.0

# LED control for ReSpeaker (optional)
pixel-ring>=0.1.0

# UUID generation
# (built-in, no package needed)

# OpenTelemetry (upgraded to 1.28.x for protobuf 5.x compatibility)
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2

# for local VAD
onnxruntime>=1.16.0

